Centralize parameter declaration invariants
Amp-Thread-ID: https://ampcode.com/threads/T-01a01db9-5c33-758a-a219-97fd03d21ff0
